Fix getting stuck in wall bug
Due to java doubles being somewhat more "precise" and bedrock positions being sent in floats instead, this caused bedrock players to get stuck in blocks when jumping or running near them in a certain way, thus causing the server to try and correct their position, potentially flagging anticheats and causing the server to print a "moved wrongly" message in console.

See: https://stackoverflow.com/questions/322749/retain-precision-with-double-in-java

+++ connector/src/main/java/org/geysermc/connector/network/translators/java/entity/player/JavaPlayerPositionRotationTranslator.java
@@ -87,10 +87,9 @@ public class JavaPlayerPositionRotationTranslator extends PacketTranslator<Serve
         if (!packet.getRelative().isEmpty()) {
             entity.moveRelative(session, packet.getX(), packet.getY() + EntityType.PLAYER.getOffset() + 0.1f, packet.getZ(), packet.getYaw(), packet.getPitch(), true);
         } else {
-            float xDis = Math.abs(entity.getPosition().getX() - (float) packet.getX());
-            float yDis = entity.getPosition().getY() - (float) packet.getY();
-            float zDis = Math.abs(entity.getPosition().getZ() - (float) packet.getZ());
-
+            double xDis = Math.abs(entity.getPosition().getX() - packet.getX());
+            double yDis = entity.getPosition().getY() - packet.getY();
+            double zDis = Math.abs(entity.getPosition().getZ() - packet.getZ());
             if (xDis > 1.5 || (yDis < 1.45 || yDis > (session.isJumping() ? 4.3 : (session.isSprinting() ? 2.5 : 1.9))) || zDis > 1.5) {
                 entity.moveAbsolute(session, Vector3f.from(packet.getX(), packet.getY(), packet.getZ()), packet.getYaw(), packet.getPitch(), true);
             }
